+++ plugins/module_utils/__init__.py
@@ -1,2 +1,0 @@
-# #!/usr/bin/python
-# Copyright 2021 Dell Inc. or its subsidiaries. All Rights Reserved